Refactor traversal to allow forking into more thread tasks
Depending on N and FILTER_DEPTH, thousands or millions of thread tasks
can be created.  There is a memory cost associated with each task, but
it should be on the order of 10s or 100s of megabytes for each million
thread tasks, so it doesn't seem like a big limitation even for large
values of N (N>20).

On my home machine, an M1 Mac mini, initializing a million thread tasks
took a little over a minute.  So again, even when initializing many
thread tasks for large values of N, the overhead is minimal.

Added "checkpoints" which are intended to allow resuming
partially-complete calculations, but they are not working reliably.